# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/courses/collections/list/mine": {
            "get": {
                "description": "获取用户收藏的课程列表，可以分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "我的收藏列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前收藏项的ID，用于分页",
                        "name": "cur_collection_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页数量限制",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回收藏列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web.CollectedCourseVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/courses/list/mine": {
            "get": {
                "description": "获取用户的课程列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "我的课程列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "年份，格式为YYYY",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "学期，如1、2、3",
                        "name": "term",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web.ProfileCourseVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/courses/{courseId}/collect": {
            "post": {
                "description": "根据课程ID，收藏或取消收藏指定课程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "课程收藏和取消",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "收藏或取消收藏的请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CourseCollectReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/courses/{courseId}/detail": {
            "get": {
                "description": "根据课程ID获取课程详情信息,包括成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "获取课程详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.PublicCourseVo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/courses/{courseId}/tags": {
            "get": {
                "description": "包括课程特点和考核方式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "获取课程标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.CourseTagsVo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ginx.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码，非 0 表示失败",
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "description": "错误或成功 描述",
                    "type": "string"
                }
            }
        },
        "web.CollectedCourseVo": {
            "type": "object",
            "properties": {
                "collection_id": {
                    "type": "integer"
                },
                "composite_score": {
                    "type": "number"
                },
                "credit": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_collected": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "school": {
                    "type": "string"
                },
                "teacher": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "web.CourseCollectReq": {
            "type": "object",
            "properties": {
                "collect": {
                    "type": "boolean"
                }
            }
        },
        "web.CourseTagsVo": {
            "type": "object",
            "properties": {
                "assessments": {
                    "description": "标签:数量",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "features": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "web.Grade": {
            "type": "object",
            "properties": {
                "final": {
                    "type": "number"
                },
                "regular": {
                    "type": "number"
                },
                "term": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "year": {
                    "type": "string"
                }
            }
        },
...
        "web.ProfileCourseVo": {
            "type": "object",
            "properties": {
                "evaluated": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "teacher": {
                    "type": "string"
                },
                "term": {
                    "description": "学年，1/2/3",
                    "type": "string"
                },
                "year": {
                    "description": "学期，2018",
                    "type": "string"
                }
            }
        },
        "web.PublicCourseVo": {
            "type": "object",
            "properties": {
                "assessments": {
                    "description": "标签:数量",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "composite_score": {
                    "type": "number"
                },
                "credit": {
                    "type": "number"
                },
                "features": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "grades": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.Grade"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_collected": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "school": {
                    "type": "string"
                },
                "teacher": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}